requirements.txt
markdown
requests
Websockets
python-slugify
nostr
Bech32
pytz
